+++ packages/cosmwasm-stargate/src/modules/wasm/aminomessages.spec.ts
@@ -10,6 +10,7 @@ import {
   MsgStoreCode,
   MsgUpdateAdmin,
 } from "cosmjs-types/cosmwasm/wasm/v1/tx";
+import { AccessType } from "cosmjs-types/cosmwasm/wasm/v1/types";
 import Long from "long";
 
 import {
@@ -40,6 +41,36 @@ describe("AminoTypes", () => {
         value: {
           sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
           wasm_byte_code: "WUVMTE9XIFNVQk1BUklORQ==",
+          instantiate_permission: undefined,
+        },
+      };
+      expect(aminoMsg).toEqual(expected);
+    });
+
+    it("works for MsgStoreCode with access type", () => {
+      const msg: MsgStoreCode = {
+        sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        wasmByteCode: fromBase64("WUVMTE9XIFNVQk1BUklORQ=="),
+        instantiatePermission: {
+          permission: AccessType.ACCESS_TYPE_ONLY_ADDRESS,
+          address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          addresses: [],
+        },
+      };
+      const aminoMsg = new AminoTypes(createWasmAminoConverters()).toAmino({
+        typeUrl: "/cosmwasm.wasm.v1.MsgStoreCode",
+        value: msg,
+      });
+      const expected: AminoMsgStoreCode = {
+        type: "wasm/MsgStoreCode",
+        value: {
+          sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          wasm_byte_code: "WUVMTE9XIFNVQk1BUklORQ==",
+          instantiate_permission: {
+            permission: "OnlyAddress",
+            address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+            addresses: undefined,
+          },
         },
       };
       expect(aminoMsg).toEqual(expected);
@@ -270,6 +301,35 @@ describe("AminoTypes", () => {
         sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
         wasmByteCode: fromBase64("WUVMTE9XIFNVQk1BUklORQ=="),
         instantiatePermission: undefined,
+      };
+      expect(msg).toEqual({
+        typeUrl: "/cosmwasm.wasm.v1.MsgStoreCode",
+        value: expectedValue,
+      });
+    });
+
+    it("works for MsgStoreCode with access type", () => {
+      const aminoMsg: AminoMsgStoreCode = {
+        type: "wasm/MsgStoreCode",
+        value: {
+          sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          wasm_byte_code: "WUVMTE9XIFNVQk1BUklORQ==",
+          instantiate_permission: {
+            permission: "OnlyAddress",
+            address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+            addresses: undefined,
+          },
+        },
+      };
+      const msg = new AminoTypes(createWasmAminoConverters()).fromAmino(aminoMsg);
+      const expectedValue: MsgStoreCode = {
+        sender: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+        wasmByteCode: fromBase64("WUVMTE9XIFNVQk1BUklORQ=="),
+        instantiatePermission: {
+          permission: AccessType.ACCESS_TYPE_ONLY_ADDRESS,
+          address: "cosmos1pkptre7fdkl6gfrzlesjjvhxhlc3r4gmmk8rs6",
+          addresses: [],
+        },
       };
       expect(msg).toEqual({
         typeUrl: "/cosmwasm.wasm.v1.MsgStoreCode",