+++ packages/cosmwasm-stargate/src/signingcosmwasmclient.ts
@@ -51,6 +51,7 @@ import {
   MsgStoreCode,
   MsgUpdateAdmin,
 } from "cosmjs-types/cosmwasm/wasm/v1/tx";
+import { AccessConfig } from "cosmjs-types/cosmwasm/wasm/v1/types";
 import Long from "long";
 import pako from "pako";
 
@@ -284,6 +285,7 @@ export class SigningCosmWasmClient extends CosmWasmClient {
     wasmCode: Uint8Array,
     fee: StdFee | "auto" | number,
     memo = "",
+    instantiatePermission?: AccessConfig,
   ): Promise<UploadResult> {
     const compressed = pako.gzip(wasmCode, { level: 9 });
     const storeCodeMsg: MsgStoreCodeEncodeObject = {
@@ -291,6 +293,7 @@ export class SigningCosmWasmClient extends CosmWasmClient {
       value: MsgStoreCode.fromPartial({
         sender: senderAddress,
         wasmByteCode: compressed,
+        instantiatePermission,
       }),
     };